Generalised the Driver class so that it won't need to be modified when we add more controllers / actions, and established a convention for naming of controller classes.

+++ controller/AdministrationController.php
@@ -13,26 +13,32 @@ class AdministrationController {
 	/**
 	 * Displays the main admin page.
 	 */
-	public function display() : void {
-		$users = new Users ();
-		$_SESSION ['users'] = $users->getUsers ();
-		
-		if (isset ( $_POST ['add'] )) {
-			unset ( $_POST ['add'] );
-			$this->add();
-		}
-		
-		if (isset ( $_POST ['update'] )) {
-			unset ( $_POST ['update'] );
-			$this->update();
-		}
-		
-		if (isset ( $_POST ['changePassword'] )) {
-			unset ( $_POST ['changePassword'] );
-			$this->changePassword();
-		}
+	public function index() : void {
 
-		include 'view/layout/administration.php';
+		if ($this->auth())
+		{
+			$users = new Users ();
+			$_SESSION ['users'] = $users->getUsers ();
+
+			if (isset ( $_POST ['add'] )) {
+				unset ( $_POST ['add'] );
+				$this->add();
+			}
+
+			if (isset ( $_POST ['update'] )) {
+				unset ( $_POST ['update'] );
+				$this->update();
+			}
+
+			if (isset ( $_POST ['changePassword'] )) {
+				unset ( $_POST ['changePassword'] );
+				$this->changePassword();
+			}
+
+			include 'view/layout/administration.php';
+		} else {
+			header('Location: ' . PATH . 'Home');
+		}
 	}
 
 	/**
@@ -40,11 +46,12 @@ class AdministrationController {
 	 */
 	public function add(): void
 	{
-		$user = new User ();
-		$v = new Validation ();
+		if ($this->auth()) {
+			$user = new User ();
+			$v = new Validation ();
 
-		// Process on submission of user.
-		if (isset ($_POST ['user'])) {
+			// Process on submission of user.
+			if (isset ($_POST ['user'])) {
 
 			// Validate the user.
 			try {
@@ -113,6 +120,9 @@ class AdministrationController {
 				}
 			}
 		}
+		} else {
+			header('Location: ' . PATH . 'Home');
+		}
 	}
 
 	/**
@@ -120,48 +130,52 @@ class AdministrationController {
 	 */
 	public function update(): void
 	{
-		$user = new User ();
-		$v = new Validation ();
+		if ($this->auth()) {
+			$user = new User ();
+			$v = new Validation ();
 
-		// Process on submission of user.
-		if (isset ($_POST ['user'])) {
+			// Process on submission of user.
+			if (isset ($_POST ['user'])) {
 
-			// Validate the user.
-			try {
-				$v->userName($_POST ['user']);
-			} catch (ValidationException $e) {
-				$_SESSION ['error'] = $e->getError();
-			}
-
-			if (isset ($_SESSION ['error'])) {
-				unset ($_POST ['user']);
-				header('Location: Administration');
-			} else {
-				$user->user = $_POST ['user'];
-				unset ($_POST ['user']);
-
-				// Validate the email.
+				// Validate the user.
 				try {
-					$v->email($_POST ['email']);
+					$v->userName($_POST ['user']);
 				} catch (ValidationException $e) {
 					$_SESSION ['error'] = $e->getError();
 				}
 
 				if (isset ($_SESSION ['error'])) {
-					unset ($_POST ['email']);
+					unset ($_POST ['user']);
 					header('Location: Administration');
 				} else {
-					$user->email = $_POST ['email'];
-					unset ($_POST ['email']);
-					$user->status = $_POST ['status'];
-					unset ($_POST ['status']);
-					$user->userID = $_POST ['userID'];
-					unset ($_POST ['userID']);
-					$user->update();
+					$user->user = $_POST ['user'];
+					unset ($_POST ['user']);
 
-					header('Location: ' . PATH . 'Administration');
+					// Validate the email.
+					try {
+						$v->email($_POST ['email']);
+					} catch (ValidationException $e) {
+						$_SESSION ['error'] = $e->getError();
+					}
+
+					if (isset ($_SESSION ['error'])) {
+						unset ($_POST ['email']);
+						header('Location: Administration');
+					} else {
+						$user->email = $_POST ['email'];
+						unset ($_POST ['email']);
+						$user->status = $_POST ['status'];
+						unset ($_POST ['status']);
+						$user->userID = $_POST ['userID'];
+						unset ($_POST ['userID']);
+						$user->update();
+
+						header('Location: ' . PATH . 'Administration');
+					}
 				}
 			}
+		} else {
+			header('Location: ' . PATH . 'Home');
 		}
 	}
 
@@ -170,27 +184,15 @@ class AdministrationController {
 	 */
 	public function changePassword(): void
 	{
-		$user = new User ();
-		$v = new Validation ();
+		if ($this->auth()) {
+			$user = new User ();
+			$v = new Validation ();
 
-		if (isset ($_POST ['password']) && isset ($_POST ['confirm'])) {
+			if (isset ($_POST ['password']) && isset ($_POST ['confirm'])) {
 
-			// Validate the password.
-			try {
-				$v->password($_POST ['password']);
-			} catch (ValidationException $e) {
-				$_SESSION ['error'] = $e->getError();
-			}
-
-			if (isset ($_SESSION ['error'])) {
-				unset ($_POST ['password']);
-				unset ($_POST ['confirm']);
-				header('Location: Administration');
-			} else {
-
-				// Compare passwords.
+				// Validate the password.
 				try {
-					$v->comparePasswords($_POST ['password'], $_POST ['confirm']);
+					$v->password($_POST ['password']);
 				} catch (ValidationException $e) {
 					$_SESSION ['error'] = $e->getError();
 				}
@@ -200,16 +202,32 @@ class AdministrationController {
 					unset ($_POST ['confirm']);
 					header('Location: Administration');
 				} else {
-					$user->password = $_POST ['password'];
-					unset ($_POST ['password']);
-					unset ($_POST ['confirm']);
-					$user->userID = $_POST ['userID'];
-					unset ($_POST ['userID']);
-					$user->update();
 
-					header('Location: ' . PATH . 'Administration');
+					// Compare passwords.
+					try {
+						$v->comparePasswords($_POST ['password'], $_POST ['confirm']);
+					} catch (ValidationException $e) {
+						$_SESSION ['error'] = $e->getError();
+					}
+
+					if (isset ($_SESSION ['error'])) {
+						unset ($_POST ['password']);
+						unset ($_POST ['confirm']);
+						header('Location: Administration');
+					} else {
+						$user->password = $_POST ['password'];
+						unset ($_POST ['password']);
+						unset ($_POST ['confirm']);
+						$user->userID = $_POST ['userID'];
+						unset ($_POST ['userID']);
+						$user->update();
+
+						header('Location: ' . PATH . 'Administration');
+					}
 				}
 			}
+		} else {
+			header('Location: ' . PATH . 'Home');
 		}
 	}
 
@@ -218,19 +236,23 @@ class AdministrationController {
 	 */
 	public function edit($id): void
 	{
-		$user = new User ();
-		$user->userID = $id;
+		if ($this->auth()) {
+			$user = new User ();
+			$user->userID = $id;
 
-		if ($user->get()) {
-			$_SESSION ['editUser'] = [
-				'userID' => $user->userID,
-				'user' => $user->user,
-				'email' => $user->email,
-				'status' => $user->status
-			];
+			if ($user->get()) {
+				$_SESSION ['editUser'] = [
+					'userID' => $user->userID,
+					'user' => $user->user,
+					'email' => $user->email,
+					'status' => $user->status
+				];
+			}
+
+			include 'view/layout/editUser.php';
+		} else {
+			header('Location: ' . PATH . 'Home');
 		}
-
-		include 'view/layout/editUser.php';
 	}
 
 	/**
@@ -238,9 +260,27 @@ class AdministrationController {
 	 */
 	public function delete($id): void
 	{
-		$user = new User ();
-		$user->userID = $id;
-		$user->delete();
-		header('Location: ' . PATH . 'Administration');
+		if ($this->auth()) {
+			$user = new User ();
+			$user->userID = $id;
+			$user->delete();
+			header('Location: ' . PATH . 'Administration');
+		} else {
+			header('Location: ' . PATH . 'Home');
+		}
 	}
+
+
+	private function auth(){
+		if(isset($_SESSION[MODULE]) && isset($_SESSION['userID'])){
+			if($_SESSION['userID'] > 0){
+				return true;
+			} else {
+				return false;
+			}
+		} else {
+			return false;
+		}
+	}
+
 }
